chore: move limits to a separate file
